adding user semester model
